// src/types/SpaceflightArticle.ts
import { Launch } from './Launch';
import { Event } from './Event';

export interface SpaceflightArticle {
  id: number;
  title: string;
  url: string;
  image_url: string;
  news_site: string;
  summary: string;
  published_at: string;
  updated_at: string;
  featured: boolean;
  launches: Launch[];
  events: Event[];
}